refactor(query): move date helpers into timestamp functions
