+++ src/query/expression/src/utils/display.rs
@@ -296,7 +296,7 @@ impl Display for ScalarRef<'_> {
             ScalarRef::String(s) => write!(f, "{}", QuotedString(s, '\'')),
             ScalarRef::Timestamp(t) => write!(f, "'{}'", timestamp_to_string(*t, &TimeZone::UTC)),
             ScalarRef::TimestampTz(t) => write!(f, "'{}'", t),
-            ScalarRef::Date(d) => write!(f, "'{}'", date_to_string(*d as i64, &TimeZone::UTC)),
+            ScalarRef::Date(d) => write!(f, "'{}'", date_to_string(*d as i64)),
             ScalarRef::Interval(interval) => write!(f, "'{}'", interval_to_string(interval)),
             ScalarRef::Array(col) => write!(f, "[{}]", col.iter().join(", ")),
             ScalarRef::Map(col) => {
@@ -368,7 +368,7 @@ pub fn scalar_ref_to_string(value: &ScalarRef) -> String {
     match value {
         ScalarRef::String(s) => s.to_string(),
         ScalarRef::Timestamp(t) => format!("{}", timestamp_to_string(*t, &TimeZone::UTC)),
-        ScalarRef::Date(d) => format!("{}", date_to_string(*d as i64, &TimeZone::UTC)),
+        ScalarRef::Date(d) => format!("{}", date_to_string(*d as i64)),
         ScalarRef::Interval(interval) => format!("{}", interval_to_string(interval)),
         ScalarRef::Bitmap(bits) => {
             let rb = deserialize_bitmap(bits).unwrap();
