+++ src/query/expression/src/utils/mod.rs
@@ -19,7 +19,6 @@ pub mod bitmap;
 pub mod block_debug;
 pub mod block_thresholds;
 mod column_from;
-pub mod date_helper;
 pub mod display;
 pub mod filter_helper;
 pub mod serialize;
