+++ src/query/formats/src/output_format/json.rs
@@ -15,10 +15,11 @@
 use databend_common_expression::DataBlock;
 use databend_common_expression::ScalarRef;
 use databend_common_expression::TableSchemaRef;
-use databend_common_expression::date_helper::DateConverter;
 use databend_common_expression::types::VectorScalarRef;
+use databend_common_expression::types::date::date_from_days;
 use databend_common_expression::types::interval::interval_to_string;
 use databend_common_expression::types::number::NumberScalar;
+use databend_common_expression::types::timestamp::timestamp_from_micros;
 use databend_common_io::deserialize_bitmap;
 use databend_common_io::prelude::OutputFormatSettings;
 use geozero::ToJson;
@@ -94,14 +95,14 @@ fn scalar_to_json(
         }),
         ScalarRef::Decimal(x) => Ok(serde_json::to_value(x.to_string()).unwrap()),
         ScalarRef::Date(v) => {
-            let dt = DateConverter::to_date(&v, &format.jiff_timezone);
+            let dt = date_from_days(v);
             Ok(serde_json::to_value(strtime::format("%Y-%m-%d", dt).unwrap()).unwrap())
         }
         ScalarRef::Interval(v) => {
             Ok(serde_json::to_value(interval_to_string(&v).to_string()).unwrap())
         }
         ScalarRef::Timestamp(v) => {
-            let dt = DateConverter::to_timestamp(&v, &format.jiff_timezone);
+            let dt = timestamp_from_micros(v, &format.jiff_timezone);
             Ok(serde_json::to_value(strtime::format("%Y-%m-%d %H:%M:%S", &dt).unwrap()).unwrap())
         }
         ScalarRef::TimestampTz(v) => Ok(serde_json::to_value(v.to_string()).unwrap()),
