+++ src/query/functions/benches/bench.rs
@@ -318,12 +318,12 @@ mod datetime_fast_path {
     use databend_common_expression::Evaluator;
     use databend_common_expression::Expr;
     use databend_common_expression::FunctionContext;
-    use databend_common_expression::date_helper::DateConverter;
     use databend_common_expression::type_check;
     use databend_common_expression::types::DataType;
     use databend_common_expression::types::string::StringColumn;
     use databend_common_expression::types::string::StringColumnBuilder;
     use databend_common_expression::types::timestamp::microseconds_to_days;
+    use databend_common_expression::types::timestamp::timestamp_from_micros;
     use databend_common_expression::types::timestamp::timestamp_to_string;
     use databend_common_expression_test_support as parser;
     use databend_common_functions::BUILTIN_FUNCTIONS;
@@ -360,7 +360,7 @@ mod datetime_fast_path {
                 let formatted = timestamp_to_string(micros, &tz_sh).to_string();
                 string_builder.put_and_commit(formatted);
 
-                let zoned = micros.to_timestamp(&tz_sh);
+                let zoned = timestamp_from_micros(micros, &tz_sh);
                 let offset_secs = zoned.offset().seconds();
                 let offset_hours = offset_secs / 3600;
                 let offset_minutes = (offset_secs.abs() % 3600) / 60;
