+++ src/query/pipeline/transforms/src/processors/transforms/window/transform_window.rs
@@ -32,8 +32,6 @@ use databend_common_expression::Scalar;
 use databend_common_expression::ScalarRef;
 use databend_common_expression::SortColumnDescription;
 use databend_common_expression::arithmetics_type::ResultTypeOfUnary;
-use databend_common_expression::date_helper::EvalMonthsImpl;
-use databend_common_expression::date_helper::calc_date_to_timestamp;
 use databend_common_expression::types::AccessType;
 use databend_common_expression::types::DataType;
 use databend_common_expression::types::DateType;
@@ -44,6 +42,8 @@ use databend_common_expression::types::NumberScalar;
 use databend_common_expression::types::NumberType;
 use databend_common_expression::types::TimestampType;
 use databend_common_expression::with_number_mapped_type;
+use databend_common_functions::scalars::dt_func::date_arithmetic::EvalMonthsImpl;
+use databend_common_functions::scalars::dt_func::date_conversion::calc_date_to_timestamp;
 use databend_common_pipeline::core::Event;
 use databend_common_pipeline::core::InputPort;
 use databend_common_pipeline::core::OutputPort;
