+++ src/query/task_support/src/system_tables/task_history.rs
@@ -28,7 +28,6 @@ use databend_common_expression::DataBlock;
 use databend_common_expression::FromData;
 use databend_common_expression::FunctionContext;
 use databend_common_expression::Scalar;
-use databend_common_expression::date_helper::DateConverter;
 use databend_common_expression::expr::*;
 use databend_common_expression::filter_helper::FilterHelpers;
 use databend_common_expression::infer_table_schema;
@@ -39,6 +38,7 @@ use databend_common_expression::types::StringType;
 use databend_common_expression::types::TimestampType;
 use databend_common_expression::types::UInt64Type;
 use databend_common_expression::types::VariantType;
+use databend_common_expression::types::timestamp::timestamp_from_micros;
 use databend_common_functions::BUILTIN_FUNCTIONS;
 use databend_common_meta_app::schema::TableIdent;
 use databend_common_meta_app::schema::TableInfo;
@@ -162,16 +162,22 @@ impl AsyncSystemTable for TaskHistoryTable {
                 if col_name == "scheduled_time"
                     && let Scalar::Timestamp(s) = scalar
                 {
-                    scheduled_time_end =
-                        Some(s.to_timestamp(&TimeZone::UTC).timestamp().to_string());
+                    scheduled_time_end = Some(
+                        timestamp_from_micros(*s, &TimeZone::UTC)
+                            .timestamp()
+                            .to_string(),
+                    );
                 }
             });
             find_gt_filter(&expr, &mut |col_name, scalar| {
                 if col_name == "scheduled_time"
                     && let Scalar::Timestamp(s) = scalar
                 {
-                    scheduled_time_start =
-                        Some(s.to_timestamp(&TimeZone::UTC).timestamp().to_string());
+                    scheduled_time_start = Some(
+                        timestamp_from_micros(*s, &TimeZone::UTC)
+                            .timestamp()
+                            .to_string(),
+                    );
                 }
             });
         }
